# src/swagger/swagger.json
{
  "info": {
    "description": "API documentation for Boiler",
    "title": "Swagger",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "paths": {
    "/user": {
      "post": {
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUser"
            }
          }
        ],
        "description": "",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/OPStatus"
            }
          }
        }
      },
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "type": "integer",
            "description": "Limit"
          },
          {
            "in": "query",
            "name": "offset",
            "type": "integer",
            "description": "Offset"
          },
          {
            "in": "query",
            "name": "name",
            "type": "string",
            "description": "Name contains"
          }
        ],
        "description": "",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      },
      "put": {
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "description": "",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/OPStatus"
            }
          }
        }
      }
    },
    "/friend": {
      "post": {
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateFriend"
            }
          }
        ],
        "description": "",
        "tags": [
          "Friend"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/OPStatus"
            }
          }
        }
      },
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "type": "string",
            "format": "uuid",
            "required": true,
            "description": "User ID"
          },
          {
            "in": "query",
            "name": "level",
            "type": "integer",
            "required": true,
            "description": "Level of friendship"
          }
        ],
        "description": "",
        "tags": [
          "Friend"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "CreateUser": {
      "type": "object",
      "properties": {
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "avatar": {
          "type": "string",
          "format": "url"
        }
      }
    },
    "CreateFriend": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "string",
          "format": "uuid"
        },
        "friend_id": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "avatar": {
          "type": "string",
          "format": "url"
        }
      }
    },
    "OPStatus": {
      "type": "string"
    }
  },
  "responses": {},
  "parameters": {},
  "tags": [
    {
      "name": "User",
      "description": "To manage user data"
    },
    {
      "name": "Friend",
      "description": "To manage friend data"
    }
  ]
}